# specs/001-mvp-publicacao-lote/contracts/api.openapi.yaml
openapi: 3.1.0
info:
  title: AdPub — API interna (feature 001)
  version: 0.1.0
  description: >
    Contrato entre a UI (Next.js) e a API (Fastify). Toda rota exige sessão
    (cookie Auth.js) e é filtrada por papel + escopo de conta.
    Nenhuma rota chama a Meta para escrita de forma síncrona: escrita é enfileirada.
servers:
  - url: /api/v1
security:
  - session: []
paths:
...
  /ad-accounts:
    get:
      summary: Contas visíveis ao usuário
      parameters:
        - { name: client_id, in: query, schema: { type: string, format: uuid } }
      responses: { "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/AdAccount" } } } } } }
  /ad-accounts/{id}:
    patch:
      summary: Define padrões da conta (página, IG, pixel, cliente, teto/dia)
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/AdAccountDefaults" } } } }
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AdAccount" } } } } }
  /ad-accounts/{id}/campaigns:
    get:
      summary: Campanhas ativas (cache) para seletor
      parameters: [{ $ref: "#/components/parameters/Id" }, { name: q, in: query, schema: { type: string } }]
      responses: { "200": { description: OK } }
  /ad-accounts/{id}/adsets:
    get:
      summary: Conjuntos ativos (cache) para seletor
      parameters: [{ $ref: "#/components/parameters/Id" }, { name: campaign_id, in: query, schema: { type: string } }]
      responses: { "200": { description: OK } }
...
  /assets:
    get:
      summary: Biblioteca do cliente
      parameters:
        - { name: client_id, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: kind, in: query, schema: { type: string, enum: [image, video] } }
        - { name: status, in: query, schema: { type: string, enum: [ok, rejected] } }
      responses: { "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Asset" } } } } } }
    post:
      summary: Upload direto (multipart) — valida e deduplica
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                client_id: { type: string, format: uuid }
                files: { type: array, items: { type: string, format: binary } }
      responses: { "201": { description: Criados/reaproveitados, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Asset" } } } } } }
  /assets/import-drive:
    post:
      summary: Enfileira importação de pasta do Google Drive
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [client_id, folder_url]
              properties:
                client_id: { type: string, format: uuid }
                folder_url: { type: string }
                recursive: { type: boolean, default: true }
      responses: { "202": { description: Enfileirado, content: { application/json: { schema: { $ref: "#/components/schemas/JobRef" } } } } }
...
components:
  securitySchemes:
    session: { type: apiKey, in: cookie, name: authjs.session-token }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string } }
...
  schemas:
    JobRef:
      type: object
      properties: { job_id: { type: string }, queue: { type: string } }
...
    AdAccount:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        currency: { type: string }
        timezone_name: { type: string }
        client_id: { type: [string, "null"] }
        default_page_id: { type: [string, "null"] }
        default_ig_user_id: { type: [string, "null"] }
        default_pixel_id: { type: [string, "null"] }
        daily_ad_cap: { type: integer }
        paused_until: { type: [string, "null"], format: date-time }
        last_synced_at: { type: string, format: date-time }
    AdAccountDefaults:
      type: object
      properties:
        client_id: { type: string }
        default_page_id: { type: string }
        default_ig_user_id: { type: string }
        default_pixel_id: { type: string }
        daily_ad_cap: { type: integer }
...
    Asset:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [image, video] }
        filename: { type: string }
        width: { type: integer }
        height: { type: integer }
        aspect_ratio: { type: string }
        duration_ms: { type: [integer, "null"] }
        size_bytes: { type: integer }
        thumbnail_url: { type: string }
        validation:
          type: object
          properties:
            status: { type: string, enum: [ok, rejected] }
            errors: { type: array, items: { type: string } }
            warnings: { type: array, items: { type: string } }
...